+++ lib/src/variant/variants/misc.dart
@@ -1,9 +1,8 @@
 part of '../variant.dart';
 
 class MiscVariants {
-  static Variant hoppelPoppel() => Variant.standard()
-      .copyWith(name: 'Hoppel-Poppel')
-      .withPieces({'N': PieceType.knibis(), 'B': PieceType.biskni()});
+  @Deprecated('Use FairyVariants.hoppelPoppel()')
+  static Variant hoppelPoppel() => FairyVariants.hoppelPoppel();
 
   static Variant spawn() => Variant.standard().copyWith(
         name: 'Spawn Chess',
@@ -141,15 +140,6 @@ class MiscVariants {
           ),
         ],
       );
-
-  static Variant grasshopper() =>
-      Variant.standard().withPiece('G', PieceType.grasshopper()).copyWith(
-            name: 'Grasshopper Chess',
-            startPosition:
-                'rnbqkbnr/gggggggg/pppppppp/8/8/PPPPPPPP/GGGGGGGG/RNBQKBNR'
-                ' w KQkq - 0 1',
-            materialConditions: MaterialConditions.none,
-          );
 
   static Variant legan() => Variant(
         name: 'Legan Chess',